+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- [Patch] Enabled possibility to create multiple ComEthernetip instances.
+
 ### Fixed
 - [Patch] Fixed shutdown behavior of ComEthernetip.
 - [Patch] Fixed log output of error string when timeout occurs (e.g. acknowlege_faults).
@@ -21,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [Patch] Increased timeout for enabling the powerstage from 1 to 5 seconds.
 - [Patch] Fixed return value of TelegramHandler: disable_powerstage().
-  
+
 ### Changed
 - [Patch] Updated logging info output strings.