+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- [Patch] Added error handling to parameter-set-load for missing parameter mappings.
+
 ### Changed
 - [Patch] Run branch pipeline only if no MR pipeline exists.
+- [Patch] Upgraded pnu_map.csv.
 
 ## v0.10.7 - 21.07.23
 ### Added